refactor: rename `helper` directory
