+++ module/applications/actor/common-sheet.mjs
@@ -1,11 +1,11 @@
 import DocumentSheetMixinEd from "../api/document-sheet-mixin.mjs";
-import { getSetting } from "../../settings.mjs";
+import { getSetting } from "../../helpers/settings.mjs";
 import TruePatternData from "../../data/thread/true-pattern.mjs";
 import PromptFactory from "../global/prompt-factory.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import DialogEd from "../api/dialog.mjs";
 import * as SYSTEM from "../../config/system.mjs";
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 const { ActorSheetV2 } = foundry.applications.sheets;
