+++ module/data/item/devotion.mjs
@@ -3,7 +3,7 @@ import IncreasableAbilityTemplate from "./templates/increasable-ability.mjs";
 import DialogEd from "../../applications/api/dialog.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as LEGEND from "../../config/legend.mjs";
-import { createContentLink } from "../../helper/formatting.mjs";
+import { createContentLink } from "../../helpers/formatting.mjs";
 
 
 /**
