+++ module/data/item/path.mjs
@@ -4,8 +4,8 @@ import DialogEd from "../../applications/api/dialog.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as LEGEND from "../../config/legend.mjs";
 import SiblingDocumentField from "../fields/sibling-document-field.mjs";
-import { getSingleGlobalItemByEdid } from "../../helper/document.mjs";
-import { createContentLink } from "../../helper/formatting.mjs";
+import { getSingleGlobalItemByEdid } from "../../helpers/document.mjs";
+import { createContentLink } from "../../helpers/formatting.mjs";
 
 /**
  * Data model template with information on path items.
