+++ module/data/item/questor.mjs
@@ -7,8 +7,8 @@ import * as LEGEND from "../../config/legend.mjs";
 import * as DOCUMENT_DATA from "../../config/document-data.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import SiblingDocumentField from "../fields/sibling-document-field.mjs";
-import { getSingleGlobalItemByEdid } from "../../helper/document.mjs";
-import { createContentLink } from "../../helper/formatting.mjs";
+import { getSingleGlobalItemByEdid } from "../../helpers/document.mjs";
+import { createContentLink } from "../../helpers/formatting.mjs";
 
 const { isEmpty } = foundry.utils;
