+++ module/data/other/character-generation.mjs
@@ -1,4 +1,3 @@
-
 import NamegiverTemplate from "../actor/templates/namegiver.mjs";
 import MappingField from "../fields/mapping-field.mjs";
 import SparseDataModel from "../abstract/sparse-data-model.mjs";
@@ -7,7 +6,7 @@ import * as ACTORS from "../../config/actors.mjs";
 import * as LEGEND from "../../config/legend.mjs";
 import { getAttributeStepFromValue } from "../../utils/earthdawn.mjs";
 import { sum } from "../../utils/math.mjs";
-import { getSingleGlobalItemByEdid, prefixKeysForDeletion } from "../../helper/document.mjs";
+import { getSingleGlobalItemByEdid, prefixKeysForDeletion } from "../../helpers/document.mjs";
 import { filterObject, mapObject } from "../../utils/object.mjs";
