+++ module/data/roll/attack.mjs
@@ -2,7 +2,7 @@ import EdRollOptions from "./common.mjs";
 import { SYSTEM_TYPES } from "../../constants/constants.mjs";
 import * as ITEMS from "../../config/items.mjs";
 import * as EFFECTS from "../../config/effects.mjs";
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 /**
  * @typedef { object } EdAttackRollOptionsInitializationData
