+++ module/data/roll/attuning.mjs
@@ -1,6 +1,6 @@
 import EdRollOptions from "./common.mjs";
 import * as MAGIC from "../../config/magic.mjs";
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 /**
  * Roll options for attuning spells to matrices or grimoires.
