+++ module/data/roll/initiative.mjs
@@ -1,6 +1,6 @@
 import EdRollOptions from "./common.mjs";
 
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 /**
  * Roll options for initiative rolls.
