+++ module/data/roll/jump-up.mjs
@@ -1,7 +1,7 @@
 import EdRollOptions from "./common.mjs";
-import { getSetting } from "../../settings.mjs";
+import { getSetting } from "../../helpers/settings.mjs";
 import * as STATUSES from "../../config/statuses.mjs";
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 /**
  * Roll options for jump up tests.
