+++ module/data/roll/knockdown.mjs
@@ -1,5 +1,5 @@
 import EdRollOptions from "./common.mjs";
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 /**
  * @typedef { object } EdKnockdownRollOptionsInitializationData
