+++ module/data/roll/weaving.mjs
@@ -1,6 +1,6 @@
 import EdRollOptions from "./common.mjs";
 import * as MAGIC from "../../config/magic.mjs";
-import { createContentAnchor } from "../../helper/formatting.mjs";
+import { createContentAnchor } from "../../helpers/formatting.mjs";
 
 /**
  * @typedef { object } EdThreadWeavingRollOptionsInitializationData
